Reorganize AWS custom waiter unit tests
